# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Descripcion del API para el submodulo de rutas",
    "version": "1.0.0",
    "title": "Ms-rutas"
  },
  "host": "100.26.181.55:80",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/nodecar/createCar": {
      "post": {
        "tags": [
          "Nodo Auto"
        ],
        "summary": "Agregar un nuevo auto en la base de datos",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto auto que va a ser guardado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idOwner": {
                  "type": "string"
                },
                "nameOwner": {
                  "type": "string"
                },
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idNodecar": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/nodecar/getCars": {
      "get": {
        "tags": [
          "Nodo Auto"
        ],
        "summary": "Consultar todos los autos en la base de datos",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "idOwner": {
                  "type": "string"
                },
                "nameOwner": {
                  "type": "string"
                },
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "No existen autos"
          }
        }
      }
    },
    "/nodecar/getIdCarbyLicense": {
      "post": {
        "tags": [
          "Nodo Auto"
        ],
        "summary": "Obtener el id en base de datos de un auto por placa ",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Placa del auto que va a ser consultada",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idNodeCar": {
                  "type": "integer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/nodecar/getCarbyLicense": {
      "post": {
        "tags": [
          "Nodo Auto"
        ],
        "summary": "Obtener un auto segun la placa",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Placa del auto que va a ser consultada",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idOwner": {
                  "type": "string"
                },
                "nameOwner": {
                  "type": "string"
                },
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/nodecar/modifyCar": {
      "put": {
        "tags": [
          "Nodo Auto"
        ],
        "summary": "Modificar un auto con la placa del auto",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto auto que va a ser modificado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idOwner": {
                  "type": "string"
                },
                "nameOwner": {
                  "type": "string"
                },
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/nodecar/deleteCar": {
      "delete": {
        "tags": [
          "Nodo Auto"
        ],
        "summary": "Eliminar un auto por la placa",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Placa del auto que va a ser eliminado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "licenseCar": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}